Refactor import paths and add Variables component
