+++ src/app/(withAuth)/courses/[courseId]/programs/[programId]/StepProblem.tsx
@@ -6,8 +6,9 @@ import { useRef } from 'react';
 import { SyntaxHighlighter } from '../../../../../../components/organisms/SyntaxHighlighter';
 import type { TurtleGraphicsHandle } from '../../../../../../components/organisms/TurtleGraphics';
 import { TurtleGraphics } from '../../../../../../components/organisms/TurtleGraphics';
-import { Variables } from '../../../../../../components/organisms/Variables';
 import { solveProblem } from '../../../../../lib/solveProblem';
+
+import { Variables } from './Variables';
 
 interface StepProblemProps {
   beforeCheckPointLine: number;
